requirements.txt
django~=5.0
djangorestframework~=3.14.0
celery~=5.3.6
requests~=2.31.0
ujson~=5.9.0
pyyaml~=6.0.1
django-rest-passwordreset>=1.3.0
redis==5.0.7
psycopg2-binary==2.9.9
drf-standardized-errors==0.14.0
django-filter==24.2
django-redis==5.4.0


